refactor: move log-common in utils
